@@ -1,15 +1,22 @@
+from ..base.rl.registration import make  # noqa F402
+from ..base.rl.registration import make_parameter  # noqa F402
+from ..base.rl.registration import make_remote_memory  # noqa F402
+from ..base.rl.registration import make_trainer  # noqa F402
+from ..base.rl.registration import make_worker  # noqa F402
+
+# from . import ql_agent57  # noqa F402
 # from . import alphazero  # noqa F402
 # from . import _muzero  # noqa F402
-from . import agent57  # noqa F402
-from . import agent57_light  # noqa F402
+# from . import dynaq  # noqa F402
+# from . import mcts  # noqa F402
+# from . import agent57  # noqa F402
+# from . import agent57_light  # noqa F402
 from . import c51  # noqa F402
 from . import dqn  # noqa F402
-from . import dynaq  # noqa F402
-from . import mcts  # noqa F402
-from . import memory  # noqa F402
 from . import ql  # noqa F402
-from . import ql_agent57  # noqa F402
 from . import r2d2  # noqa F402
 from . import rainbow  # noqa F402
-from . import sac  # noqa F402
-from .registory import make  # noqa F402
+from .rulebase import human  # noqa F402
+from .rulebase import random_play  # noqa F402
+
+# from . import sac  # noqa F402
