@@ -1,14 +1,16 @@
 import random
-from collections import deque
 from dataclasses import dataclass
 from typing import Any, List, Optional, Tuple, cast
 
 import numpy as np
 import tensorflow as tf
 import tensorflow.keras as keras
-from srl.base.rl import DiscreteActionConfig, RLParameter, RLRemoteMemory, RLTrainer, RLWorker
+from srl.base.env.env_for_rl import EnvForRL
+from srl.base.rl.algorithms.neuralnet_discrete import DiscreteActionConfig, DiscreteActionWorker
+from srl.base.rl.base import RLParameter, RLTrainer
+from srl.base.rl.registration import register
 from srl.rl.functions.model import ImageLayerType, create_input_layers
-from srl.rl.registory import register
+from srl.rl.remote_memory.experience_replay_buffer import ExperienceReplayBuffer
 from tensorflow.keras import layers as kl
 
 """
@@ -67,7 +69,24 @@ class Config(DiscreteActionConfig):
         assert self.batch_size < self.memory_warmup_size
 
 
-register(Config, __name__)
+register(
+    Config,
+    __name__ + ":RemoteMemory",
+    __name__ + ":Parameter",
+    __name__ + ":Trainer",
+    __name__ + ":Worker",
+)
+
+
+# ------------------------------------------------------
+# RemoteMemory
+# ------------------------------------------------------
+class RemoteMemory(ExperienceReplayBuffer):
+    def __init__(self, *args):
+        super().__init__(*args)
+        self.config = cast(Config, self.config)
+
+        self.init(self.config.capacity)
 
 
 # ------------------------------------------------------
@@ -124,47 +143,6 @@ class Parameter(RLParameter):
 
 
 # ------------------------------------------------------
-# RemoteMemory
-# ------------------------------------------------------
-class RemoteMemory(RLRemoteMemory):
-    def __init__(self, *args):
-        super().__init__(*args)
-        self.config = cast(Config, self.config)
-
-        self.memory = deque(maxlen=self.config.capacity)
-        self.invalid_memory = deque(maxlen=self.config.capacity)
-
-    def length(self) -> int:
-        return len(self.memory)
-
-    def restore(self, data: Any) -> None:
-        self.memory = data[0]
-        self.invalid_memory = data[1]
-
-    def backup(self):
-        return [self.memory, self.invalid_memory]
-
-    # ---------------------------
-    def length_invalid(self) -> int:
-        return len(self.invalid_memory)
-
-    def add(self, batch):
-        self.memory.append(batch)
-
-    def add_invalid(self, batch):
-        self.invalid_memory.append(batch)
-
-    def sample(self):
-        return random.sample(self.memory, self.config.batch_size)
-
-    def sample_invalid(self):
-        return random.sample(self.invalid_memory, self.config.batch_size)
-
-    def clear_invalid(self):
-        return self.invalid_memory.clear()
-
-
-# ------------------------------------------------------
 # Trainer
 # ------------------------------------------------------
 class Trainer(RLTrainer):
@@ -172,7 +150,7 @@ class Trainer(RLTrainer):
         super().__init__(*args)
         self.config = cast(Config, self.config)
         self.parameter = cast(Parameter, self.parameter)
-        self.memory = cast(RemoteMemory, self.memory)
+        self.remote_memory = cast(RemoteMemory, self.remote_memory)
 
         self.optimizer = keras.optimizers.Adam(learning_rate=self.config.lr)
         self.loss = keras.losses.Huber()
@@ -184,17 +162,11 @@ class Trainer(RLTrainer):
 
     def train(self):
 
-        if self.memory.length() < self.config.memory_warmup_size:
+        if self.remote_memory.length() < self.config.memory_warmup_size:
             return {}
 
-        batchs = self.memory.sample()
+        batchs = self.remote_memory.sample(self.config.batch_size)
         loss = self._train_on_batchs(batchs)
-
-        # invalid action
-        mem_invalid_len = self.memory.length_invalid()
-        if mem_invalid_len > self.config.memory_warmup_size:
-            batchs = self.memory.sample_invalid()
-            self._train_on_batchs(batchs)
 
         # targetと同期
         if self.train_count % self.config.target_model_update_interval == 0:
@@ -203,7 +175,6 @@ class Trainer(RLTrainer):
         self.train_count += 1
         return {
             "loss": loss,
-            "mem2": mem_invalid_len,
         }
 
     def _train_on_batchs(self, batchs):
@@ -211,14 +182,11 @@ class Trainer(RLTrainer):
         states = []
         actions = []
         n_states = []
-        rewards = []
-        dones = []
+        next_invalid_actions = []
         for b in batchs:
             states.append(b["states"][:-1])
             actions.append(b["action"])
             n_states.append(b["states"][1:])
-            rewards.append(b["reward"])
-            dones.append(b["done"])
         states = np.asarray(states)
         n_states = np.asarray(n_states)
 
@@ -227,20 +195,26 @@ class Trainer(RLTrainer):
         n_q_target = self.parameter.q_target(n_states).numpy()
 
         # 各バッチのQ値を計算
-        target_q = []
-        for i in range(len(rewards)):
-            if dones[i]:
-                gain = rewards[i]
+        target_q = np.zeros(len(batchs))
+        for i, b in enumerate(batchs):
+            reward = b["reward"]
+            done = b["done"]
+            next_invalid_actions = b["next_invalid_actions"]
+            if done:
+                gain = reward
             else:
                 # DoubleDQN: indexはonlineQから選び、値はtargetQを選ぶ
                 if self.config.enable_double_dqn:
+                    n_q[i] = [(-np.inf if a in next_invalid_actions else v) for a, v in enumerate(n_q[i])]
                     n_act_idx = np.argmax(n_q[i])
                 else:
+                    n_q_target[i] = [
+                        (-np.inf if a in next_invalid_actions else v) for a, v in enumerate(n_q_target[i])
+                    ]
                     n_act_idx = np.argmax(n_q_target[i])
                 maxq = n_q_target[i][n_act_idx]
-                gain = rewards[i] + self.config.gamma * maxq
-            target_q.append(gain)
-        target_q = np.asarray(target_q)
+                gain = reward + self.config.gamma * maxq
+            target_q[i] = gain
 
         with tf.GradientTape() as tape:
             q = self.parameter.q_online(states)
@@ -260,15 +234,14 @@ class Trainer(RLTrainer):
 # ------------------------------------------------------
 # Worker
 # ------------------------------------------------------
-class Worker(RLWorker):
+class Worker(DiscreteActionWorker):
     def __init__(self, *args):
         super().__init__(*args)
         self.config = cast(Config, self.config)
         self.parameter = cast(Parameter, self.parameter)
-        self.memory = cast(RemoteMemory, self.memory)
+        self.remote_memory = cast(RemoteMemory, self.remote_memory)
 
         self.dummy_state = np.full(self.config.env_observation_shape, self.config.dummy_state_val)
-        self.invalid_action_reward = -1
         self.step = 0
 
         if self.config.exploration_steps > 0:
@@ -278,13 +251,13 @@ class Worker(RLWorker):
             ) / self.config.exploration_steps
             self.final_epsilon = self.config.final_epsilon
 
-    def on_reset(self, state: np.ndarray, valid_actions: List[int], _) -> None:
+    def call_on_reset(self, state: np.ndarray, invalid_actions: List[int]) -> None:
         self.recent_states = [self.dummy_state for _ in range(self.config.window_length + 1)]
 
         self.recent_states.pop(0)
         self.recent_states.append(state)
 
-    def policy(self, _state: np.ndarray, valid_actions: List[int], _) -> Tuple[int, Any]:
+    def call_policy(self, state: np.ndarray, invalid_actions: List[int]) -> Tuple[int, Any]:
 
         if self.training:
             if self.config.exploration_steps > 0:
@@ -299,31 +272,29 @@ class Worker(RLWorker):
 
         if random.random() < epsilon:
             # epsilonより低いならランダム
-            action = random.choice(valid_actions)
+            action = random.choice([a for a in range(self.config.nb_actions) if a not in invalid_actions])
         else:
             state = self.recent_states[1:]
             q = self.parameter.q_online(np.asarray([state]))[0].numpy()
 
-            # valid actions以外は -inf にする
-            q = np.array([(v if i in valid_actions else -np.inf) for i, v in enumerate(q)])
+            # invalid actionsは -inf にする
+            q = [(-np.inf if a in invalid_actions else v) for a, v in enumerate(q)]
 
             # 最大値を選ぶ（複数はほぼないので無視）
             action = int(np.argmax(q))
 
         return action, action
 
-    def on_step(
+    def call_on_step(
         self,
         state: np.ndarray,
         action: Any,
         next_state: np.ndarray,
         reward: float,
         done: bool,
-        valid_actions: List[int],
-        next_valid_actions: List[int],
-        _,
+        invalid_actions: List[int],
+        next_invalid_actions: List[int],
     ):
-
         self.recent_states.pop(0)
         self.recent_states.append(next_state)
 
@@ -343,36 +314,23 @@ class Worker(RLWorker):
             "action": action,
             "reward": reward,
             "done": done,
+            "next_invalid_actions": next_invalid_actions,
         }
-        self.memory.add(batch)
-
-        # --- invalid action
-        if self.invalid_action_reward > reward - 1:
-            self.invalid_action_reward = reward - 1
-            self.memory.clear_invalid()
-
-        states = self.recent_states[:]
-        states[-1] = self.dummy_state
-        for a in range(self.config.nb_actions):
-            if a in valid_actions:
-                continue
-
-            batch = {
-                "states": states,
-                "action": a,
-                "reward": self.invalid_action_reward,
-                "done": True,
-            }
-            self.memory.add_invalid(batch)
+        self.remote_memory.add(batch)
 
         return {}
 
-    def render(self, state_: np.ndarray, valid_actions: List[int], action_to_str) -> None:
+    def render(
+        self,
+        state: np.ndarray,
+        invalid_actions: List[int],
+        env: EnvForRL,
+    ):
         state = self.recent_states[1:]
         q = self.parameter.q_online(np.asarray([state]))[0].numpy()
         maxa = np.argmax(q)
         for a in range(self.config.nb_actions):
-            if a not in valid_actions:
+            if a in invalid_actions:
                 s = "x"
             else:
                 s = " "
@@ -380,7 +338,7 @@ class Worker(RLWorker):
                 s += "*"
             else:
                 s += " "
-            s += f"{action_to_str(a)}: {q[a]:5.3f}"
+            s += f"{env.action_to_str(a)}: {q[a]:6.3f}"
             print(s)
 
 
