@@ -1,15 +1,23 @@
 import tensorflow as tf
 from tensorflow.keras import layers as kl
+import tensorflow_addons as tfa
 
 
-def create_dueling_network_layers(c, nb_actions: int, dense_units: int, dueling_type: str):
+def create_dueling_network_layers(
+    c, nb_actions: int, dense_units: int, dueling_type: str, enable_noisy_dense: bool = False
+):
+    if enable_noisy_dense:
+        Dense = tfa.layers.NoisyDense
+    else:
+        Dense = kl.Dense
+
     # value
-    v = kl.Dense(dense_units, activation="relu", kernel_initializer="he_normal")(c)
-    v = kl.Dense(1, kernel_initializer="truncated_normal", name="v")(v)
+    v = Dense(dense_units, activation="relu", kernel_initializer="he_normal")(c)
+    v = Dense(1, kernel_initializer="truncated_normal", name="v")(v)
 
     # advance
-    adv = kl.Dense(dense_units, activation="relu", kernel_initializer="he_normal")(c)
-    adv = kl.Dense(nb_actions, kernel_initializer="truncated_normal", name="adv")(adv)
+    adv = Dense(dense_units, activation="relu", kernel_initializer="he_normal")(c)
+    adv = Dense(nb_actions, kernel_initializer="truncated_normal", name="adv")(adv)
 
     # 連結で結合
     c = kl.Concatenate()([v, adv])
@@ -24,7 +32,7 @@ def create_dueling_network_layers(c, nb_actions: int, dense_units: int, dueling_
             output_shape=(nb_actions,),
         )(c)
     elif dueling_type == "":  # naive
-        c = kl.Lambda(lambda a: tf.expand_dims(a[:, 0], -1) + a[:, 1:], output_shape=(config.nb_actions,))(c)
+        c = kl.Lambda(lambda a: tf.expand_dims(a[:, 0], -1) + a[:, 1:], output_shape=(nb_actions,))(c)
     else:
         raise ValueError("dueling_network_type is undefined")
 
