@@ -1,7 +1,4 @@
-from .action_continuous_processor import ActionContinuousProcessor  # noqa F402
-from .action_discrete_processor import ActionDiscreteProcessor  # noqa F402
+from .continuous_processor import ContinuousProcessor  # noqa F402
+from .discrete_processor import DiscreteProcessor  # noqa F402
 from .image_processor import ImageProcessor  # noqa F402
-from .observation_continuous_processor import \
-    ObservationContinuousProcessor  # noqa F402
-from .observation_discrete_processor import \
-    ObservationDiscreteProcessor  # noqa F402
+from .observation_box_processor import ObservationBoxProcessor  # noqa F402
