@@ -1,17 +1,18 @@
 import random
-from collections import deque
 from dataclasses import dataclass
 from typing import Any, List, Tuple, cast
 
 import numpy as np
 import tensorflow as tf
 import tensorflow.keras as keras
-from srl.base.rl import DiscreteActionConfig, RLParameter, RLRemoteMemory, RLTrainer, RLWorker
+from srl.base.env.env_for_rl import EnvForRL
+from srl.base.rl.algorithms.neuralnet_discrete import DiscreteActionConfig, DiscreteActionWorker
+from srl.base.rl.base import RLParameter, RLTrainer
+from srl.base.rl.registration import register
 from srl.rl.functions.common import calc_epsilon_greedy_probs, inverse_rescaling, random_choice_by_probs, rescaling
 from srl.rl.functions.dueling_network import create_dueling_network_layers
 from srl.rl.functions.model import ImageLayerType, create_input_layers_lstm_stateful
-from srl.rl.memory import factory
-from srl.rl.registory import register
+from srl.rl.remote_memory.priority_experience_replay import PriorityExperienceReplay
 from tensorflow.keras import layers as kl
 
 """
@@ -68,7 +69,6 @@ class Config(DiscreteActionConfig):
     # DuelingNetwork
     enable_dueling_network: bool = True
     dueling_network_type: str = "average"
-    dueling_dense_units: int = 512
 
     # Priority Experience Replay
     capacity: int = 100_000
@@ -92,7 +92,29 @@ class Config(DiscreteActionConfig):
         assert self.batch_size < self.memory_warmup_size
 
 
-register(Config, __name__)
+register(
+    Config,
+    __name__ + ":RemoteMemory",
+    __name__ + ":Parameter",
+    __name__ + ":Trainer",
+    __name__ + ":Worker",
+)
+
+# ------------------------------------------------------
+# RemoteMemory
+# ------------------------------------------------------
+class RemoteMemory(PriorityExperienceReplay):
+    def __init__(self, *args):
+        super().__init__(*args)
+        self.config = cast(Config, self.config)
+
+        self.init(
+            self.config.memory_name,
+            self.config.capacity,
+            self.config.memory_alpha,
+            self.config.memory_beta_initial,
+            self.config.memory_beta_steps,
+        )
 
 
 # ------------------------------------------------------
@@ -116,9 +138,7 @@ class _QNetwork(keras.Model):
         c = kl.LSTM(config.lstm_units, stateful=True, name="lstm")(c)
 
         if config.enable_dueling_network:
-            c = create_dueling_network_layers(
-                c, config.nb_actions, config.dueling_dense_units, config.dueling_network_type
-            )
+            c = create_dueling_network_layers(c, config.nb_actions, config.dense_units, config.dueling_network_type)
         else:
             c = kl.Dense(config.dense_units, activation="relu", kernel_initializer="he_normal")(c)
             c = kl.Dense(config.nb_actions, kernel_initializer="truncated_normal")(c)
@@ -167,60 +187,6 @@ class Parameter(RLParameter):
 
 
 # ------------------------------------------------------
-# RemoteMemory
-# ------------------------------------------------------
-class RemoteMemory(RLRemoteMemory):
-    def __init__(self, *args):
-        super().__init__(*args)
-        self.config = cast(Config, self.config)
-
-        self.memory = factory.create(
-            self.config.memory_name,
-            {
-                "capacity": self.config.capacity,
-                "alpha": self.config.memory_alpha,
-                "beta_initial": self.config.memory_beta_initial,
-                "beta_steps": self.config.memory_beta_steps,
-            },
-        )
-        self.invalid_memory = deque(maxlen=self.config.capacity)
-
-    def length(self) -> int:
-        return len(self.memory)
-
-    def restore(self, data: Any) -> None:
-        self.memory.restore(data[0])
-        self.invalid_memory = data[1]
-
-    def backup(self):
-        d = [self.memory.backup(), self.invalid_memory]
-        return d
-
-    # ---------------------------
-
-    def add(self, batch, priority):
-        self.memory.add(batch, priority)
-
-    def sample(self, step: int) -> Tuple[list, list, list]:
-        return self.memory.sample(self.config.batch_size, step)
-
-    def update(self, indexes: List[int], batchs: List[Any], priorities: List[float]) -> None:
-        self.memory.update(indexes, batchs, priorities)
-
-    def length_invalid(self) -> int:
-        return len(self.invalid_memory)
-
-    def add_invalid(self, batch):
-        self.invalid_memory.append(batch)
-
-    def sample_invalid(self):
-        return random.sample(self.invalid_memory, self.config.batch_size)
-
-    def clear_invalid(self):
-        return self.invalid_memory.clear()
-
-
-# ------------------------------------------------------
 # Trainer
 # ------------------------------------------------------
 class Trainer(RLTrainer):
@@ -228,7 +194,7 @@ class Trainer(RLTrainer):
         super().__init__(*args)
         self.config = cast(Config, self.config)
         self.parameter = cast(Parameter, self.parameter)
-        self.memory = cast(RemoteMemory, self.memory)
+        self.remote_memory = cast(RemoteMemory, self.remote_memory)
 
         self.optimizer = keras.optimizers.Adam(learning_rate=self.config.lr)
         self.loss = keras.losses.Huber()
@@ -240,19 +206,13 @@ class Trainer(RLTrainer):
 
     def train(self):
 
-        if self.memory.length() < self.config.memory_warmup_size:
+        if self.remote_memory.length() < self.config.memory_warmup_size:
             return {}
 
-        indexes, batchs, weights = self.memory.sample(self.train_count)
+        indexes, batchs, weights = self.remote_memory.sample(self.train_count, self.config.batch_size)
         td_errors, loss = self._train_on_batchs(batchs, weights)
         priorities = np.abs(td_errors) + 0.0001
-        self.memory.update(indexes, batchs, priorities)
-
-        # invalid action
-        mem_invalid_len = self.memory.length_invalid()
-        if mem_invalid_len > self.config.memory_warmup_size:
-            batchs = self.memory.sample_invalid()
-            self._train_on_batchs(batchs, [1 for _ in range(self.config.batch_size)])
+        self.remote_memory.update(indexes, batchs, priorities)
 
         # targetと同期
         if self.train_count % self.config.target_model_update_interval == 0:
@@ -261,7 +221,6 @@ class Trainer(RLTrainer):
         self.train_count += 1
         return {
             "loss": loss,
-            "mem2": mem_invalid_len,
             "priority": np.mean(priorities),
         }
 
@@ -315,7 +274,7 @@ class Trainer(RLTrainer):
         )
         return td_error, loss
 
-    # Q値(LSTM hidden states)の予測はforward、td_error,retraceはbackで予測する必要あり
+    # Q値(LSTM hidden states)の予測はforward、td_error,retraceはbackwardで予測する必要あり
     def _train_steps(
         self,
         states_list,
@@ -409,23 +368,23 @@ class Trainer(RLTrainer):
 # ------------------------------------------------------
 # Worker
 # ------------------------------------------------------
-class Worker(RLWorker):
+class Worker(DiscreteActionWorker):
     def __init__(self, *args):
         super().__init__(*args)
         self.config = cast(Config, self.config)
         self.parameter = cast(Parameter, self.parameter)
-        self.memory = cast(RemoteMemory, self.memory)
+        self.remote_memory = cast(RemoteMemory, self.remote_memory)
 
         self.dummy_state = np.full(self.config.env_observation_shape, self.config.dummy_state_val, dtype=np.float32)
         self.invalid_action_reward = -1
 
-    def on_reset(self, state: np.ndarray, valid_actions: List[int], _) -> None:
+    def call_on_reset(self, state: np.ndarray, invalid_actions: List[int]) -> None:
         self.recent_states = [self.dummy_state for _ in range(self.config.burnin + self.config.multisteps + 1)]
         self.recent_actions = [random.randint(0, self.config.nb_actions - 1) for _ in range(self.config.multisteps)]
         self.recent_probs = [1.0 / self.config.nb_actions for _ in range(self.config.multisteps)]
         self.recent_rewards = [0.0 for _ in range(self.config.multisteps)]
         self.recent_done = [False for _ in range(self.config.multisteps)]
-        self.recent_valid_actions = [[] for _ in range(self.config.multisteps + 1)]
+        self.recent_invalid_actions = [[] for _ in range(self.config.multisteps + 1)]
 
         self.hidden_state = self.parameter.q_online.init_hidden_state()
         self.recent_hidden_states = [
@@ -435,10 +394,10 @@ class Worker(RLWorker):
 
         self.recent_states.pop(0)
         self.recent_states.append(state.astype(np.float32))
-        self.recent_valid_actions.pop(0)
-        self.recent_valid_actions.append(valid_actions)
+        self.recent_invalid_actions.pop(0)
+        self.recent_invalid_actions.append(invalid_actions)
 
-    def policy(self, state: np.ndarray, valid_actions: List[int], _) -> Tuple[int, Any]:
+    def call_policy(self, state: np.ndarray, invalid_actions: List[int]) -> Tuple[int, Any]:
         state = np.asarray([[state]] * self.config.batch_size)
         q, self.hidden_state = self.parameter.q_online(state, self.hidden_state)
         q = q[0].numpy()
@@ -448,21 +407,20 @@ class Worker(RLWorker):
         else:
             epsilon = self.config.test_epsilon
 
-        probs = calc_epsilon_greedy_probs(q, valid_actions, epsilon, self.config.nb_actions)
+        probs = calc_epsilon_greedy_probs(q, invalid_actions, epsilon, self.config.nb_actions)
         action = random_choice_by_probs(probs)
 
         return action, (action, probs[action], q[action])
 
-    def on_step(
+    def call_on_step(
         self,
         state: np.ndarray,
         action_: Any,
         next_state: np.ndarray,
         reward: float,
         done: bool,
-        valid_actions: List[int],
-        next_valid_actions: List[int],
-        _,
+        invalid_actions: List[int],
+        next_invalid_actions: List[int],
     ):
         if not self.training:
             return {}
@@ -470,10 +428,6 @@ class Worker(RLWorker):
         action = action_[0]
         prob = action_[1]
         q = action_[2]
-
-        if self.invalid_action_reward > reward - 1:
-            self.invalid_action_reward = reward - 1
-            self.memory.clear_invalid()
 
         self.recent_states.pop(0)
         self.recent_states.append(next_state.astype(np.float32))
@@ -485,8 +439,8 @@ class Worker(RLWorker):
         self.recent_rewards.append(reward)
         self.recent_done.pop(0)
         self.recent_done.append(done)
-        self.recent_valid_actions.pop(0)
-        self.recent_valid_actions.append(next_valid_actions)
+        self.recent_invalid_actions.pop(0)
+        self.recent_invalid_actions.append(next_invalid_actions)
         self.recent_hidden_states.pop(0)
         self.recent_hidden_states.append(
             [
@@ -510,39 +464,11 @@ class Worker(RLWorker):
                 self.recent_rewards.append(0.0)
                 self.recent_done.pop(0)
                 self.recent_done.append(True)
-                self.recent_valid_actions.pop(0)
-                self.recent_valid_actions.append([])
+                self.recent_invalid_actions.pop(0)
+                self.recent_invalid_actions.append([])
                 self.recent_hidden_states.pop(0)
 
                 self._add_memory(q, priority)
-
-        # --- valid action
-        if False:  # TODO
-            states = self.recent_states[:]
-            states[-1] = self.dummy_state
-            bundle_states = self.recent_bundle_states[:]
-            bundle_states[-1] = states
-            rewards = self.recent_rewards[:]
-            rewards[-1] = self.invalid_action_reward
-            probs = self.recent_probs[:]
-            probs[-1] = 1.0
-            dones = self.recent_done[:]
-            dones[-1] = True
-            for a in range(self.config.nb_actions):
-                if a in valid_actions:
-                    continue
-                actions = self.recent_actions[:]
-                actions[-1] = a
-
-                batch = {
-                    "states": bundle_states,
-                    "actions": actions,
-                    "probs": probs,
-                    "rewards": rewards,
-                    "dones": dones,
-                    "valid_actions": self.recent_valid_actions[:],
-                }
-                self.memory.add_invalid(batch)
 
         return {"priority": priority}
 
@@ -554,7 +480,7 @@ class Worker(RLWorker):
             "probs": self.recent_probs[:],
             "rewards": self.recent_rewards[:],
             "dones": self.recent_done[:],
-            "valid_actions": self.recent_valid_actions[:],
+            "invalid_actions": self.recent_invalid_actions[:],
             "burnin_states": self.recent_states[: self.config.burnin],
             "hidden_states": self.recent_hidden_states[0],
         }
@@ -569,17 +495,22 @@ class Worker(RLWorker):
                 # target_q = self.parameter.calc_target_q([batch])[0]
                 # priority = abs(target_q - q) + 0.0001
 
-        self.memory.add(batch, priority)
+        self.remote_memory.add(batch, priority)
         return priority
 
-    def render(self, state: np.ndarray, valid_actions: List[int], action_to_str) -> None:
+    def render(
+        self,
+        state: np.ndarray,
+        invalid_actions: List[int],
+        env: EnvForRL,
+    ) -> None:
         state = np.asarray([[state]] * self.config.batch_size)
         q, self.hidden_state = self.parameter.q_online(state, self.hidden_state)
         q = q[0].numpy()
 
         maxa = np.argmax(q)
         for a in range(self.config.nb_actions):
-            if a not in valid_actions:
+            if a not in invalid_actions:
                 s = "x"
             else:
                 s = " "
@@ -587,7 +518,7 @@ class Worker(RLWorker):
                 s += "*"
             else:
                 s += " "
-            s += f"{action_to_str(a)}: {q[a]:5.3f}"
+            s += f"{env.action_to_str(a)}: {q[a]:5.3f}"
             print(s)
 
 
