@@ -12,12 +12,17 @@ class Test(unittest.TestCase):
         rl_config = rl.ql.Config()
         self.tester.play_test(self, rl_config)
 
-    def test_verify(self):
+    def test_verify_grid(self):
         rl_config = rl.ql.Config(epsilon=0.5, lr=0.01)
-        self.tester.play_verify(self, "Grid-v0", rl_config, 50_000, 1000)
-        self.tester.check_action_values(self)
-        self.tester.check_policy(self)
+        self.tester.play_verify_singleplay(self, "Grid", rl_config, 50_000, 1000)
+        self.tester.verify_grid_action_values(self)
+        self.tester.verify_grid_policy(self)
+
+    def test_verify_ox(self):
+        rl_config = rl.ql.Config(epsilon=0.5, lr=0.1)
+        self.tester.play_verify_2play(self, "OX", rl_config, 100_000, 1000)
 
 
 if __name__ == "__main__":
-    unittest.main(module=__name__, defaultTest="Test.test_verify", verbosity=2)
+    # unittest.main(module=__name__, defaultTest="Test.test_play", verbosity=2)
+    unittest.main(module=__name__, defaultTest="Test.test_verify_ox", verbosity=2)
